auth fixes
