+++ nodemon.json
@@ -1,4 +1,4 @@
 {
    "verbose": true,
-   "watch": ["src"]
+   "watch": ["dist"]
 }
